CI: don't rebuild everything twice
Reverts 051a64c4b1

Gasp, I only now noticed this; it should absolutely be avoided, in general (the
Docker image is expected to contain a recent repo), and in particular after
switch creation.

Switch to a newer image instead (or just remove the date constraint in FROM but
stable builds are good).