CI: when a same-name branch exists on catala-examples, use that
This offers a workaround when a PR on Catala needs a change to the examples or
french-law: you would have the CI broken either on the PR or on master.

With this, you can ensure that the CI passes without breaking master, and then
merge the branch on catala-examples/french-law simultaneously to the PR on
Catala.