Fix handling of the lexer context
The lexing context uses a global reference. This patch ensures it is handled
carefully for the lifetime of lexing a given file.

Before that, very weird bugs could happen, for example when a file ended with an
unclosed verbatim block, the parsing of module interfaces done later on would
start in verbatim context and fail to parse e.g. module use definitions.

The patch additionally adds checks and warns about unclosed blocks at EOF, which
before that would be silently accepted (with the above pitfall).