Add list folds to the surface syntax
Temporary syntax:
```
for a initially 0 and then (a + x) for x among lst
```

and in French:
```
pour a initialement 0 puis (a + x) pour x parmi lst
```

to be elaborated with the Syntax Committee of course. Two new keywords:
`initially` and `and then` (it could have been the existing `then` keyword for
english, relying on the ambiguity of the word, but that's not portable across
languages).
Then added to the cheat-sheet and docs.

Note that detuplification takes place, making it possible both to fold multiple
lists in parallel, and to use multiple accumulators. See the tests for examples.