C backend: workaround bulk freeing of constants
Constants may need initialisation and allocations, which we do lazily, but that
doesn't combine well with the way our allocator works, freeing everything in
bulk.

This patch adds a switch to the allocator that make it use raw `malloc`, which
is turned on within the initialisation functions of constants. A cleaner
solution will be to let Catala proceed normally but do a copy to persistent
memory afterwards; we'll need to implement this copy anyway for consumers to
more easily recover results from calls to catala.