Proper support for handling #[test] runs
Instead of being treated specially in each backend, tests declared with the
`#[test]` attribute are now provisionned as soon as dcalc, with proper (empty)
arguments even for context-function variables. Then they go through the expected
transformations along with the rest of the program (including e.g. closure
conversion) and provide the body of a `main` function for the backends.