Make all test scopes implicitely public
As well as their type dependencies, because of the previous patch.

Note that this is very convenient, but without any warning, it might lead to
confusing scenarios where e.g. removing a test would made some code non longer
compile because it was making direct use of a type that was private (by
mistake?), but implicitely promoted to public because of the test...