Fix some odoc warnings
This doesn't fix everything, but given the stability of `odoc` that seems to be
a wild goose chase... On the latest (3.1) release we have errors such as:

```
Caught unexpected exception when expanding type declaration (Odoc_xref2.Expand_tools.Clash)
Failed to list assoc a
Caught unexpected exception when expanding type declaration (Failure("bah"))
```

And, for example, `@raise` doesn't tolerate Stdlib exceptions. But the results
seem fine.

Closes #375