Require at least a digit after decimal separators
This is a fix for #724 intended for discussion: after some thought I find this
solution the most elegant. In any case, I am inclined to think that `14.0` is
clearly readable to everyone, while `14.` is not a standard mathematical
notation and might be confusing.

We still need to separate tuple elements with a space, but it's interesting to
see that this is the same in French and English: `(3,142)` is a single integer
in the English syntax, and a single decimal in French. `(3, 142)` — with this
patch — is a pair of integers for both, but it used to be a syntax error in
French, requiring `(3 , 142)` instead.

This may break some existing programs which used e.g. `3.` to mean the decimal
3, though.