Fix implicit uses of the Stdlib in submodules
NOTE: this is a temporary fix, it will only load the stdlib of the root module
and apply that to all module, instead of picking the correct stdlib for each
module.

This means that using a `catala_en` module from a `catala_fr` one will break (as
soon as the used module uses the stdlib).