OCaml: use constant constructors for constructors with empty payload
This is a more targeted fix for #913 than what #914 proposes: only the OCaml
backend is affected on the specific "constructor with no payload" case.

It is mostly orthogonal though: if there is more reason to change the
representation in the compiler, #914 remains valid.

The hard part was to adjust the dark magic in the live (native
runtime)/(interpreter) translations to account for the change in the OCaml data
representation.